@@ -1,3 +1,5 @@
+import { Field } from "formik";
+
 interface RadioComProps {
   label: string;
   name: string;
@@ -19,14 +21,15 @@ const RadioField: React.FC<RadioComProps> = ({
         <div className="space-y-4 sm:flex sm:items-center sm:space-x-10 sm:space-y-0">
           {options.map(({ label, value }, index) => (
             <div className="flex items-center" key={index}>
-              <input
-                id={`${name}-${value}`}
+              <Field
+                id={`${name}-${label}`}
                 name={name}
                 type="radio"
+                value={value}
                 className="h-4 w-4 border-gray-300 text-indigo-600 focus:ring-indigo-600"
               />
               <label
-                htmlFor={`${name}-${value}`}
+                htmlFor={`${name}-${label}`}
                 className="ml-3 block text-sm font-medium leading-6 text-gray-900"
               >
                 {label}
