@@ -1,3 +1,4 @@
+import { Field } from "formik";
 import { FaCircleInfo } from "react-icons/fa6";
 
 interface TextFieldProps {
@@ -24,13 +25,13 @@ const TextField: React.FC<TextFieldProps> = ({
         {label}
       </label>
       <div className="mt-2">
-        <textarea
+        <Field
+          as="textarea"
           rows={4}
           name={name}
           id={name}
           placeholder={placeholder}
-          className="block w-full rounded-md border-0 px-3 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 resize-none"
-          defaultValue={""}
+          className="block w-full rounded-md border-0 px-3 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-green-600 sm:text-sm sm:leading-6 resize-none outline-none"
         />
       </div>
       {note && (
