@@ -37,7 +37,7 @@ const navigation = {
 
 const Footer: React.FC = () => {
   return (
-    <footer className="bg-white">
+    <footer className="">
       <div className="mx-auto max-w-7xl overflow-hidden px-6 py-20 sm:py-24 lg:px-8">
         <nav
           className="-mb-6 columns-2 sm:flex sm:justify-center sm:space-x-12"
